Revert "ARM: Change dmb domain for bionic_atomic_barrier()"
Well the original change is indeed correct. But there are devices
which are doing weird things in their proprietaries across gpu and other memory areas
(toro and toroplus as example). This ends up in serious issues on this devices
which are doing it obviosuly wrong.

Whatever till we know which devices are all effected and we can ifdef them
we revert it for now.

This reverts commit f8caadb87dfc4f2eed86462fa6fb0cab5262d52e.

Change-Id: I93ba5dff94ebdef4629d7d9f005037b190e76ad4